api/requirements.txt
fastapi
uvicorn
pyodbc
python-dotenv
pydantic
